backend/requirements.txt
fastapi
uvicorn
groq
supabase
python-dotenv
pdfplumber
pydantic
python-multipart
